Move comopnents to auth folder
