+++ src/components/auth/Registration.vue
