+++ src/router/index.js
@@ -1,8 +1,8 @@
 import Vue from 'vue'
 import Router from 'vue-router'
 import Home from '@/components/home/Home'
-import Login from '@/components/home/Login'
-import Registration from '@/components/home/Registration'
+import Login from '@/components/auth/Login'
+import Registration from '@/components/auth/Registration'
 import Campaign from '@/components/home/Campaign'
 import CampaignDetail from '@/components/home/CampaignDetail'
